todolist: add() works OK.
Gets data from cmd line & appends it to the data file.

Note:
Use this pattern to determine if a date is included in the ADD
command:
  \d\d/
This pattern might cause problems if no date is included, and the
task starts w/ \d\d/.  I think this is not a bad first solution.

Use the code in check_date() that was made for list().  This seems
to work OK so far.